requirements.txt
aenum==3.1.15
affine==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
arcgis==2.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
attrs==24.2.0
azure-core==1.32.0
azure-storage-blob==12.24.0
babel==2.16.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.2.0
blinker==1.9.0
Bottleneck==1.4.2
branca==0.8.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cftime==1.6.4.post1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
cytoolz==1.0.1
dask==2024.8.0
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distributed==2024.8.0
earthpy==0.9.4
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
fiona==1.10.1
flake8==7.1.1
folium==0.19.2
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.10.0
future==1.0.0
GDAL==3.6.4
geo==0.1.1
GeoAlchemy2==0.16.0
geomet==1.1.0
geopandas==1.0.1
geoserver-rest==2.9.1
greenlet==3.1.1
h11==0.14.0
h3==4.1.2
h5py==3.12.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
intel-cmplr-lib-ur==2025.0.4
intel-openmp==2025.0.4
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==7.8.5
isodate==0.7.2
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jdcal==1.4.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==6.1.12
jupyter-console==6.4.2
jupyter-contrib-core==0.4.2
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server_fileid==0.9.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==1.1.11
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
locket==1.0.0
lxml==5.3.0
lz4==4.3.3
mapclassify==2.8.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
mkl==2025.0.1
mkl-service==2.4.2
mkl_fft==1.3.11
mkl_random==1.2.8
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msrest==0.7.1
munch==4.0.0
munkres==1.1.4
mypy-extensions==1.0.0
nbclassic==1.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.2.1
nose==1.3.7
notebook==6.5.7
notebook_shim==0.2.4
ntlm-auth==1.5.0
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openpyxl==3.1.5
osmnx==2.0.1
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==1.0.1
pefile==2024.8.26
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.1.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
pygeos==0.14
Pygments==2.18.0
PyJWT==2.10.1
pylerc==4.0
pyodbc==5.2.0
pyogrio==0.10.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
pyspnego==0.11.2
pytest==8.3.4
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.0
python-multipart==0.0.19
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.3
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
Rtree==1.3.0
saspy==5.101.0
scikit-image==0.25.1
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
sentinelhub==3.11.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sspilib==0.2.0
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
swat==1.15.0
sympy==1.13.3
terminado==0.18.1
testpath==0.6.0
threadpoolctl==3.5.0
tifffile==2025.1.10
timezonefinder==6.5.7
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
umf==0.9.1
uri-template==1.3.0
urllib3==2.2.3
utm==0.8.0
uvicorn==0.33.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
win-inet-pton==1.1.0
wrapt==1.17.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.14.2
y-py==0.6.2
ypy-websocket==0.8.4
zict==3.0.0
